Extract general version utilities
