+++ libs/wire-api/src/Wire/API/Routes/Version.hs
@@ -39,20 +39,12 @@ where
 
 import Control.Lens ((?~))
 import Data.Aeson (FromJSON, ToJSON (..))
-import qualified Data.Aeson as Aeson
-import qualified Data.CaseInsensitive as CI
 import Data.Domain
-import Data.Metrics.Servant
 import Data.Schema
 import Data.Singletons.TH
 import qualified Data.Swagger as S
-import qualified Data.Text as Text
-import qualified Data.Text.Read as Text
 import Imports
-import qualified Network.Wai as Wai
 import Servant
-import Servant.Server.Internal.Delayed
-import Servant.Server.Internal.DelayedIO
 import Servant.Swagger
 import Wire.API.Routes.Named
 import Wire.API.VersionInfo
@@ -68,18 +60,6 @@ instance ToSchema Version where
         element 1 V1,
         element 2 V2
       ]
-
-readVersionNumber :: Text -> Maybe Integer
-readVersionNumber v = do
-  ('v', rest) <- Text.uncons v
-  case Text.decimal rest of
-    Right (n, "") -> pure n
-    _ -> Nothing
-
-mkVersion :: Integer -> Maybe Version
-mkVersion n = case Aeson.fromJSON (Aeson.Number (fromIntegral n)) of
-  Aeson.Error _ -> Nothing
-  Aeson.Success v -> pure v
 
 supportedVersions :: [Version]
 supportedVersions = [minBound .. maxBound]
@@ -117,68 +97,4 @@ type VersionAPI =
 versionSwagger :: S.Swagger
 versionSwagger = toSwagger (Proxy @VersionAPI)
 
-versionHeader :: CI.CI ByteString
-versionHeader = "X-Wire-API-Version"
-
 $(genSingletons [''Version])
-
---------------------------------------------------------------------------------
--- Servant combinators
-
-data Until (v :: Version)
-
-data From (v :: Version)
-
-instance (SingI v, HasServer api ctx) => HasServer (Until v :> api) ctx where
-  type ServerT (Until v :> api) m = ServerT api m
-
-  route _ ctx action =
-    route (Proxy @api) ctx $
-      fmap const action `addHeaderCheck` withRequest headerCheck
-    where
-      headerCheck :: Wai.Request -> DelayedIO ()
-      headerCheck req = do
-        let v =
-              toEnum $
-                maybe
-                  0
-                  (either (const 0) id . parseHeader)
-                  (lookup versionHeader (Wai.requestHeaders req))
-        when (v >= demote @v) $
-          delayedFail err404
-
-  hoistServerWithContext _ ctx f =
-    hoistServerWithContext (Proxy @api) ctx f
-
-instance HasSwagger api => HasSwagger (Until v :> api) where
-  toSwagger _ = mempty
-
-instance RoutesToPaths api => RoutesToPaths (Until v :> api) where
-  getRoutes = getRoutes @api
-
-instance (SingI v, HasServer api ctx) => HasServer (From v :> api) ctx where
-  type ServerT (From v :> api) m = ServerT api m
-
-  route _ ctx action =
-    route (Proxy @api) ctx $
-      fmap const action `addHeaderCheck` withRequest headerCheck
-    where
-      headerCheck :: Wai.Request -> DelayedIO ()
-      headerCheck req = do
-        let v =
-              toEnum $
-                maybe
-                  0
-                  (either (const 0) id . parseHeader)
-                  (lookup versionHeader (Wai.requestHeaders req))
-        when (v < demote @v) $
-          delayedFail err404
-
-  hoistServerWithContext _ ctx f =
-    hoistServerWithContext (Proxy @api) ctx f
-
-instance HasSwagger api => HasSwagger (From v :> api) where
-  toSwagger _ = toSwagger (Proxy @api)
-
-instance RoutesToPaths api => RoutesToPaths (From v :> api) where
-  getRoutes = getRoutes @api
